Task 21.1: Clean Up
This commit is all about updating the codebase to make it clearer.

To do this, the README file examples were updated to include the `INSERT`
queries which were added from the last commit. There was also some
format changes and some markdown updates, namely identifying any
shell commands as shell commands.

Another change was partitioning the existing code from `ordering.go`.
`ordering.go` had too many functions and this made it difficult to make
changes. As such, two new files were created for this sake, `writers.go`
and `helpers.go`.

`writers.go` contains the functions that are responsible for creating
the `SQL` queries that the program uses.

`helpers.go` contains other functions that are used by the `Ordering`
struct.

Finally, I got rid of the postgres URL that was accidently added to the
examples.

Other Changes:
- {POSTGRES_URL} --> "{POSTGRES_URL}" in the examples
